refactoring frontend layout
