+++ src/components/managers/ActivityManager.js
