+++ src/components/managers/AuthManager.js
