+++ src/components/managers/ClimbTypeManager.js
