+++ src/components/managers/HikeManager.js
