+++ src/components/managers/MyHikeManager.js
