+++ src/components/managers/SkillLevelManager.js
