+++ src/components/managers/TagManager.js
