+++ src/components/views/ApplicationViews.js
@@ -1,14 +1,10 @@
 import { Route, Routes } from "react-router-dom"
-import { ClimbContainer } from "../activities/climbs/ClimbContainer"
 import { ClimbEdit } from "../activities/climbs/ClimbEdit"
 import { ClimbForm } from "../activities/climbs/ClimbForm"
 import { ClimbList } from "../activities/climbs/ClimbList"
-import { ClimbSearch } from "../activities/climbs/ClimbSearch"
-import { HikeContainer } from "../activities/hikes/HikeContainer"
 import { HikeEdit } from "../activities/hikes/HikeEdit"
 import { HikeForm } from "../activities/hikes/HikeForm"
 import { HikeList } from "../activities/hikes/HikeList"
-import { HikeSearch } from "../activities/hikes/HikeSearch"
 import { Login } from "../auth/Login"
 import { Register } from "../auth/Register"
 import { Links } from "../helpfulLinks/Links"
