+++ src/index.js
@@ -1,6 +1,6 @@
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import { SummitCity } from './components/SummitCity';
+import { SummitCity } from './SummitCity';
 import './index.css';
 
 const container = document.getElementById("root")
